use 'git rev-parse' to query if subdir of repo
Previously used existence of '.git' (presumably due to 'git rev-parse'
not playing nicely with 'errexit' which is set for tests), which meant
repos weren't being included unless they were in the top level of the
repo. Resolved by using the fact errexit does not apply to the condition
of an if-statement.